calculator-console/requirements-dev.txt
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-playwright>=0.5.0
